+++ Editor/ImageHandlerRegister.cs
@@ -7,14 +7,14 @@ public static class ImageHandlerRegister
     static ImageHandlerRegister()
     {
         EditorConnection.instance.Initialize();
-        EditorConnection.instance.Register(FailedImageMessage.MessageId, ImageHandler.instance.HandleFailedImageEvent);
+        EditorConnection.instance.Register(ImageMessage.MessageId, ImageHandler.instance.HandleImageEvent);
 
         AssemblyReloadEvents.beforeAssemblyReload += Unregister;
     }
 
     private static void Unregister()
     {
-        EditorConnection.instance.Unregister(FailedImageMessage.MessageId, ImageHandler.instance.HandleFailedImageEvent);
+        EditorConnection.instance.Unregister(ImageMessage.MessageId, ImageHandler.instance.HandleImageEvent);
         AssemblyReloadEvents.beforeAssemblyReload -= Unregister;
     }
 }